// convex/functions/_generated/dataModel.d.ts
export type DataModel = {
// ...
  user: {
    document: {
      createdAt: number;
      email: string;
      emailVerified: boolean;
      image?: null | string;
      name: string;
      updatedAt: number;
      userId?: null | string;
      _id: Id<"user">;
      _creationTime: number;
    };
    fieldPaths:
      | "_creationTime"
      | "_id"
      | "createdAt"
      | "email"
      | "emailVerified"
      | "image"
      | "name"
      | "updatedAt"
      | "userId";
    indexes: {
      by_id: ["_id"];
      by_creation_time: ["_creationTime"];
      email_name: ["email", "name", "_creationTime"];
      name: ["name", "_creationTime"];
      user_email_unique: ["email", "_creationTime"];
    };
    searchIndexes: {};
    vectorIndexes: {};
  };
// ...
};

/**
 * The names of all of your Convex tables.
 */
export type TableNames = TableNamesInDataModel<DataModel>;
// ...
export type Id<TableName extends TableNames | SystemTableNames> =
  GenericId<TableName>;